+++ test/integration/nested_model.cc
@@ -16,6 +16,7 @@
  */
 
 #include <sstream>
+#include <fstream>
 #include <string>
 
 #include <gtest/gtest.h>
@@ -345,23 +346,25 @@ TEST(NestedModel, NestedInclude)
   ASSERT_NE(nullptr, upperLink3);
 
   // expect link poses from model2 and model3 to match raw link poses
-  // from model1 when transformed by model2 / model3 pose
-  EXPECT_EQ(baseLink2->RawPose(), model2Pose * baseLink1->RawPose());
-  EXPECT_EQ(baseLink3->RawPose(), model3Pose * baseLink1->RawPose());
-  EXPECT_EQ(lowerLink2->RawPose(), model2Pose * lowerLink1->RawPose());
-  EXPECT_EQ(lowerLink3->RawPose(), model3Pose * lowerLink1->RawPose());
-  EXPECT_EQ(upperLink2->RawPose(), model2Pose * upperLink1->RawPose());
-  EXPECT_EQ(upperLink3->RawPose(), model3Pose * upperLink1->RawPose());
-  // expect empty //pose/@relative_to
+  // from model1.
+  EXPECT_EQ(baseLink2->RawPose(), baseLink1->RawPose());
+  EXPECT_EQ(baseLink3->RawPose(), baseLink1->RawPose());
+  EXPECT_EQ(lowerLink2->RawPose(), lowerLink1->RawPose());
+  EXPECT_EQ(lowerLink3->RawPose(), lowerLink1->RawPose());
+  EXPECT_EQ(upperLink2->RawPose(), upperLink1->RawPose());
+  EXPECT_EQ(upperLink3->RawPose(), upperLink1->RawPose());
+  // expect //pose/@relative_to to contain the name of the nested model frame
+  // injected during parsing. model1 is not nested, so it's links should have
+  // empty //pose/@relative_to
   EXPECT_TRUE(baseLink1->PoseRelativeTo().empty());
-  EXPECT_TRUE(baseLink2->PoseRelativeTo().empty());
-  EXPECT_TRUE(baseLink3->PoseRelativeTo().empty());
+  EXPECT_EQ(name + "::__model__", baseLink2->PoseRelativeTo());
+  EXPECT_EQ(name + "_14::__model__", baseLink3->PoseRelativeTo());
   EXPECT_TRUE(lowerLink1->PoseRelativeTo().empty());
-  EXPECT_TRUE(lowerLink2->PoseRelativeTo().empty());
-  EXPECT_TRUE(lowerLink3->PoseRelativeTo().empty());
+  EXPECT_EQ(name + "::__model__", lowerLink2->PoseRelativeTo());
+  EXPECT_EQ(name + "_14::__model__", lowerLink3->PoseRelativeTo());
   EXPECT_TRUE(upperLink1->PoseRelativeTo().empty());
-  EXPECT_TRUE(upperLink2->PoseRelativeTo().empty());
-  EXPECT_TRUE(upperLink3->PoseRelativeTo().empty());
+  EXPECT_EQ(name + "::__model__", upperLink2->PoseRelativeTo());
+  EXPECT_EQ(name + "_14::__model__", upperLink3->PoseRelativeTo());
 
   // each model has 2 joints, and the joint names of the nested models have
   // been transformed
@@ -399,10 +402,429 @@ TEST(NestedModel, NestedInclude)
   EXPECT_EQ(ignition::math::Vector3d::UnitX, upperAxis1->Xyz());
   EXPECT_EQ(ignition::math::Vector3d::UnitX, lowerAxis2->Xyz());
   EXPECT_EQ(ignition::math::Vector3d::UnitX, upperAxis2->Xyz());
-  // expect //axis/xyz to be rotated for model3 since it came from
-  // SDFormat 1.4, which implies //axis/xyz/@expressed_in == "__model__"
-  const ignition::math::Vector3d rotatedAxis =
-      model3Pose.Rot() * ignition::math::Vector3d::UnitX;
-  EXPECT_EQ(rotatedAxis, lowerAxis3->Xyz());
-  EXPECT_EQ(rotatedAxis, upperAxis3->Xyz());
+  // For model3, expect //axis/xyz to use the nested model frame in
+  // //axis/xyz/@expressed_in since it came from SDFormat 1.4, which implies
+  // //axis/xyz/@expressed_in == "__model__" inside the nested model
+  EXPECT_EQ(ignition::math::Vector3d::UnitX, lowerAxis3->Xyz());
+  EXPECT_EQ(ignition::math::Vector3d::UnitX, upperAxis3->Xyz());
+  EXPECT_EQ(name + "_14::__model__", lowerAxis3->XyzExpressedIn());
+  EXPECT_EQ(name + "_14::__model__", upperAxis3->XyzExpressedIn());
+}
+
+//////////////////////////////////////////////////
+// Test parsing models with child models containg frames nested via <include>
+TEST(NestedModel, NestedModelWithFrames)
+{
+  const std::string name = "test_model_with_frames";
+  const std::string modelPath = std::string(PROJECT_SOURCE_PATH)
+      + "/test/integration/model/" + name;
+
+  const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
+
+  std::ostringstream stream;
+  std::string version = "1.7";
+  stream
+    << "<sdf version='" << version << "'>"
+    << "<world name='default'>"
+    << "  <model name='ParentModel'>"
+    << "    <include>"
+    << "      <uri>" + modelPath + "</uri>"
+    << "      <name>M1</name>"
+    << "      <pose>" << model1Pose << "</pose>"
+    << "    </include>"
+    << "  </model>"
+    << "</world>"
+    << "</sdf>";
+
+  sdf::SDFPtr sdfParsed(new sdf::SDF());
+  sdf::init(sdfParsed);
+  ASSERT_TRUE(sdf::readString(stream.str(), sdfParsed));
+
+  sdf::Root root;
+  sdf::Errors errors = root.Load(sdfParsed);
+  for (auto e : errors)
+    std::cout << e.Message() << std::endl;
+  EXPECT_TRUE(errors.empty());
+
+  const sdf::World *world = root.WorldByIndex(0);
+  ASSERT_NE(nullptr, world);
+
+  const sdf::Model *parentModel = world->ModelByIndex(0);
+  ASSERT_NE(nullptr, parentModel);
+
+  using ignition::math::Pose3d;
+  using ignition::math::Vector3d;
+  // Expected poses for frames, links and joints after nesting.
+  Pose3d frame1ExpPose = model1Pose * Pose3d(0, 0, 0, IGN_PI/2, 0, 0);
+  Pose3d frame2ExpPose = frame1ExpPose * Pose3d(0, 0, 0, 0, IGN_PI/4, 0);
+  Pose3d link1ExpPose = frame1ExpPose;
+  Pose3d link2ExpPose = frame1ExpPose * Pose3d(1, 0, 0, 0, 0, 0);
+  Pose3d joint1ExpPose = link1ExpPose;
+  Vector3d joint1AxisExpVector = frame2ExpPose.Rot() * Vector3d::UnitZ;
+  Vector3d joint1Axis2ExpVector = frame2ExpPose.Rot() * Vector3d::UnitX;
+
+  const auto *frame1 = parentModel->FrameByName("M1::F1");
+  ASSERT_NE(nullptr, frame1);
+  Pose3d frame1Pose;
+  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose);
+
+  const auto *frame2 = parentModel->FrameByName("M1::F2");
+  ASSERT_NE(nullptr, frame2);
+  Pose3d frame2Pose;
+  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose);
+
+  const auto *link1 = parentModel->LinkByName("M1::L1");
+  ASSERT_NE(nullptr, link1);
+  Pose3d link1Pose;
+  EXPECT_TRUE(link1->SemanticPose().Resolve(link1Pose).empty());
+  EXPECT_EQ(link1ExpPose, link1Pose);
+
+  const auto *visual1 = link1->VisualByName("V1");
+  ASSERT_NE(nullptr, visual1);
+  EXPECT_EQ("M1::F2", visual1->PoseRelativeTo());
+
+  const auto *collision1 = link1->CollisionByName("C1");
+  ASSERT_NE(nullptr, collision1);
+  EXPECT_EQ("M1::__model__", collision1->PoseRelativeTo());
+
+  const auto *link2 = parentModel->LinkByName("M1::L2");
+  ASSERT_NE(nullptr, link2);
+  Pose3d link2Pose;
+  EXPECT_TRUE(link2->SemanticPose().Resolve(link2Pose).empty());
+  EXPECT_EQ(link2ExpPose, link2Pose);
+
+  const auto *joint1 = parentModel->JointByName("M1::J1");
+  ASSERT_NE(nullptr, joint1);
+  Pose3d joint1Pose;
+  EXPECT_TRUE(joint1->SemanticPose().Resolve(joint1Pose, "__model__").empty());
+  EXPECT_EQ(joint1ExpPose, joint1Pose);
+
+  const auto joint1Axis = joint1->Axis(0);
+  ASSERT_NE(nullptr, joint1Axis);
+  Vector3d joint1AxisVector;
+  EXPECT_TRUE(joint1Axis->ResolveXyz(joint1AxisVector, "__model__").empty());
+  EXPECT_EQ(joint1AxisExpVector, joint1AxisVector);
+
+  const auto joint1Axis2 = joint1->Axis(1);
+  ASSERT_NE(nullptr, joint1Axis2);
+  Vector3d joint1Axis2Vector;
+  EXPECT_TRUE(joint1Axis2->ResolveXyz(joint1Axis2Vector, "__model__").empty());
+  EXPECT_EQ(joint1Axis2ExpVector, joint1Axis2Vector);
+}
+
+// Function to remove any element in //joint/axis that is not <xyz>
+static void removeNoneXyz(const sdf::ElementPtr &_elem)
+{
+  std::vector<sdf::ElementPtr> toRemove;
+  for (auto el = _elem->GetFirstElement(); el; el = el->GetNextElement())
+  {
+    if (el->GetName() != "xyz")
+    {
+      toRemove.push_back(el);
+    }
+  }
+
+  for (auto el : toRemove)
+  {
+    _elem->RemoveChild(el);
+  }
+}
+
+// Remove elements in world that are not model for comparison with
+// expectation. Also remove any element in //joint/axis that is not <xyz>
+void prepareForDirectComparison(sdf::ElementPtr _worldElem)
+{
+  std::vector<sdf::ElementPtr> toRemove;
+  for (auto elem = _worldElem->GetFirstElement(); elem;
+       elem = elem->GetNextElement())
+  {
+    if (elem->GetName() != "model")
+    {
+      toRemove.push_back(elem);
+    }
+    else
+    {
+      if (elem->HasElement("joint"))
+      {
+        for (auto joint = elem->GetElement("joint"); joint;
+             joint = joint->GetNextElement("joint"))
+        {
+          if (joint->HasElement("axis"))
+          {
+            removeNoneXyz(joint->GetElement("axis"));
+          }
+          if (joint->HasElement("axis2"))
+          {
+            removeNoneXyz(joint->GetElement("axis2"));
+          }
+        }
+      }
+    }
+  }
+
+  for (auto elem : toRemove)
+  {
+    _worldElem->RemoveChild(elem);
+  }
+}
+
+//////////////////////////////////////////////////
+// Test parsing models with child models containg frames nested via <include>
+// Compare parsed SDF with expected string
+TEST(NestedModel, NestedModelWithFramesDirectComparison)
+{
+  const std::string name = "test_model_with_frames";
+  const std::string modelPath = std::string(PROJECT_SOURCE_PATH)
+      + "/test/integration/model/" + name;
+
+  const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
+
+  std::ostringstream stream;
+  std::string version = "1.7";
+  stream
+    << "<sdf version='" << version << "'>"
+    << "<world name='default'>"
+    << "  <model name='ParentModel'>"
+    << "    <include>"
+    << "      <uri>" + modelPath + "</uri>"
+    << "      <name>M1</name>"
+    << "      <pose>" << model1Pose << "</pose>"
+    << "    </include>"
+    << "  </model>"
+    << "</world>"
+    << "</sdf>";
+
+  sdf::SDFPtr sdfParsed(new sdf::SDF());
+  sdf::init(sdfParsed);
+  ASSERT_TRUE(sdf::readString(stream.str(), sdfParsed));
+
+  auto worldElem = sdfParsed->Root()->GetElement("world");
+  prepareForDirectComparison(worldElem);
+
+  // Compare with expected output
+  const std::string expectedSdfPath =
+      std::string(PROJECT_SOURCE_PATH) +
+      "/test/integration/nested_model_with_frames_expected.sdf";
+  std::fstream fs;
+  fs.open(expectedSdfPath);
+  EXPECT_TRUE(fs.is_open());
+  std::stringstream expected;
+  fs >> expected.rdbuf();
+  EXPECT_EQ(expected.str(), sdfParsed->ToString());
+}
+
+//////////////////////////////////////////////////
+// Test parsing models that have two levels of nesting with child models
+// containg frames nested via <include>.
+// Compare parsed SDF with expected string
+TEST(NestedModel, TwoLevelNestedModelWithFramesDirectComparison)
+{
+  const std::string modelRootPath = std::string(PROJECT_SOURCE_PATH)
+      + "/test/integration/model/";
+  const std::string name = "test_nested_model_with_frames";
+  const std::string modelPath = modelRootPath + name;
+
+  const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
+
+  std::ostringstream stream;
+  std::string version = "1.7";
+  stream
+    << "<sdf version='" << version << "'>"
+    << "<world name='default'>"
+    << "  <model name='ParentModel'>"
+    << "    <include>"
+    << "      <uri>" + modelPath + "</uri>"
+    << "      <name>M1</name>"
+    << "      <pose>" << model1Pose << "</pose>"
+    << "    </include>"
+    << "  </model>"
+    << "</world>"
+    << "</sdf>";
+
+  sdf::SDFPtr sdfParsed(new sdf::SDF());
+  sdf::init(sdfParsed);
+
+  sdf::setFindCallback(
+      [&](const std::string &_file)
+      {
+        return modelRootPath + _file;
+      });
+
+  ASSERT_TRUE(sdf::readString(stream.str(), sdfParsed));
+
+  auto worldElem = sdfParsed->Root()->GetElement("world");
+  prepareForDirectComparison(worldElem);
+
+  // Compare with expected output
+  const std::string expectedSdfPath =
+      std::string(PROJECT_SOURCE_PATH) +
+      "/test/integration/two_level_nested_model_with_frames_expected.sdf";
+  std::fstream fs;
+  fs.open(expectedSdfPath);
+  EXPECT_TRUE(fs.is_open());
+  std::stringstream expected;
+  fs >> expected.rdbuf();
+  EXPECT_EQ(expected.str(), sdfParsed->ToString());
+}
+
+//////////////////////////////////////////////////
+// Test parsing models nested with <include> and that reference sibling frames
+// in their //include/pose/@relative_to attribute
+TEST(NestedModel, NestedModelWithSiblingFrames)
+{
+  const std::string name = "test_model_with_frames";
+  const std::string MODEL_PATH = std::string(PROJECT_SOURCE_PATH)
+      + "/test/integration/model/" + name;
+
+  const ignition::math::Pose3d testFramePose(0, 5, 0, 0, 0, -IGN_PI/2);
+  const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
+
+  std::ostringstream stream;
+  std::string version = "1.7";
+  stream
+    << "<sdf version='" << version << "'>"
+    << "<world name='default'>"
+    << "  <model name='ParentModel'>"
+    << "    <frame name='testFrame'>"
+    << "     <pose>" << testFramePose << "</pose>"
+    << "    </frame>"
+    << "    <include>"
+    << "      <uri>" + MODEL_PATH + "</uri>"
+    << "      <name>M1</name>"
+    << "      <pose relative_to='testFrame'>" << model1Pose << "</pose>"
+    << "    </include>"
+    << "  </model>"
+    << "</world>"
+    << "</sdf>";
+
+  sdf::SDFPtr sdfParsed(new sdf::SDF());
+  sdf::init(sdfParsed);
+  ASSERT_TRUE(sdf::readString(stream.str(), sdfParsed));
+
+  sdf::Root root;
+  sdf::Errors errors = root.Load(sdfParsed);
+  for (auto e : errors)
+    std::cout << e.Message() << std::endl;
+  EXPECT_TRUE(errors.empty());
+
+  const sdf::World *world = root.WorldByIndex(0);
+  ASSERT_NE(nullptr, world);
+
+  const sdf::Model *parentModel = world->ModelByIndex(0);
+  ASSERT_NE(nullptr, parentModel);
+
+  const sdf::Frame *nestedModel1Frame =
+      parentModel->FrameByName("M1::__model__");
+  ASSERT_NE(nullptr, nestedModel1Frame);
+
+  EXPECT_EQ("testFrame", nestedModel1Frame->PoseRelativeTo());
+
+  using ignition::math::Pose3d;
+  using ignition::math::Vector3d;
+  // Expected poses
+  Pose3d nestedModel1FrameExpPose = testFramePose * model1Pose;
+  Pose3d frame1ExpPose =
+      nestedModel1FrameExpPose * Pose3d(0, 0, 0, IGN_PI / 2, 0, 0);
+  Pose3d frame2ExpPose = frame1ExpPose * Pose3d(0, 0, 0, 0, IGN_PI / 4, 0);
+  Pose3d link1ExpPose = frame1ExpPose;
+  Pose3d link2ExpPose = frame1ExpPose * Pose3d(1, 0, 0, 0, 0, 0);
+  Pose3d joint1ExpPose = link1ExpPose;
+
+  Pose3d nestedModel1FramePose;
+  EXPECT_TRUE(
+      nestedModel1Frame->SemanticPose().Resolve(nestedModel1FramePose).empty());
+  EXPECT_EQ(nestedModel1FrameExpPose, nestedModel1FramePose);
+
+  const auto *frame1 = parentModel->FrameByName("M1::F1");
+  ASSERT_NE(nullptr, frame1);
+  Pose3d frame1Pose;
+  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose);
+
+  const auto *frame2 = parentModel->FrameByName("M1::F2");
+  ASSERT_NE(nullptr, frame2);
+  Pose3d frame2Pose;
+  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose);
+
+  const auto *link1 = parentModel->LinkByName("M1::L1");
+  ASSERT_NE(nullptr, link1);
+  Pose3d link1Pose;
+  EXPECT_TRUE(link1->SemanticPose().Resolve(link1Pose).empty());
+  EXPECT_EQ(link1ExpPose, link1Pose);
+
+  const auto *link2 = parentModel->LinkByName("M1::L2");
+  ASSERT_NE(nullptr, link2);
+  Pose3d link2Pose;
+  EXPECT_TRUE(link2->SemanticPose().Resolve(link2Pose).empty());
+  EXPECT_EQ(link2ExpPose, link2Pose);
+
+  const auto *joint1 = parentModel->JointByName("M1::J1");
+  ASSERT_NE(nullptr, joint1);
+  Pose3d joint1Pose;
+  EXPECT_TRUE(joint1->SemanticPose().Resolve(joint1Pose, "__model__").empty());
+  EXPECT_EQ(joint1ExpPose, joint1Pose);
+}
+
+//////////////////////////////////////////////////
+// Test parsing models with child models that contain only frames
+// and are nested via <include>
+TEST(NestedModel, NestedFrameOnlyModel)
+{
+  const std::string name = "frame_only_model";
+  const std::string MODEL_PATH = std::string(PROJECT_SOURCE_PATH)
+      + "/test/integration/model/" + name;
+
+  const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
+
+  std::ostringstream stream;
+  std::string version = "1.7";
+  stream
+    << "<sdf version='" << version << "'>"
+    << "<world name='default'>"
+    << "  <model name='ParentModel'>"
+    << "    <include>"
+    << "      <uri>" + MODEL_PATH + "</uri>"
+    << "      <name>M1</name>"
+    << "      <pose>" << model1Pose << "</pose>"
+    << "    </include>"
+    << "  </model>"
+    << "</world>"
+    << "</sdf>";
+
+  sdf::SDFPtr sdfParsed(new sdf::SDF());
+  sdf::init(sdfParsed);
+  ASSERT_TRUE(sdf::readString(stream.str(), sdfParsed));
+
+  sdf::Root root;
+  sdf::Errors errors = root.Load(sdfParsed);
+  for (auto e : errors)
+    std::cout << e.Message() << std::endl;
+  EXPECT_TRUE(errors.empty());
+
+  const sdf::World *world = root.WorldByIndex(0);
+  ASSERT_NE(nullptr, world);
+
+  const sdf::Model *parentModel = world->ModelByIndex(0);
+  ASSERT_NE(nullptr, parentModel);
+
+  using ignition::math::Pose3d;
+  using ignition::math::Vector3d;
+  // Expected poses for frames after nesting.
+  Pose3d frame1ExpPose = model1Pose * Pose3d(0, 0, 0, IGN_PI, 0, 0);
+  Pose3d frame2ExpPose = frame1ExpPose * Pose3d(0, 0, 0, 0, IGN_PI, 0);
+
+  const auto *frame1 = parentModel->FrameByName("M1::F1");
+  ASSERT_NE(nullptr, frame1);
+  Pose3d frame1Pose;
+  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose);
+
+  const auto *frame2 = parentModel->FrameByName("M1::F2");
+  ASSERT_NE(nullptr, frame2);
+  Pose3d frame2Pose;
+  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose);
 }